+++ src/Craft.php
@@ -10,6 +10,7 @@ use craft\behaviors\CustomFieldBehavior;
 use craft\db\Query;
 use craft\db\Table;
 use craft\helpers\App;
+use craft\helpers\ArrayHelper;
 use craft\helpers\Component;
 use craft\helpers\FileHelper;
 use GuzzleHttp\Client;
@@ -344,11 +345,8 @@ EOD;
         // Grab the config from config/guzzle.php that is used on every Guzzle request.
         $guzzleConfig = static::$app->getConfig()->getConfigFromFile('guzzle');
 
-        // Merge default into guzzle config.
-        $guzzleConfig = array_replace_recursive($guzzleConfig, $defaultConfig);
-
-        // Maybe they want to set some config options specifically for this request.
-        $guzzleConfig = array_replace_recursive($guzzleConfig, $config);
+        // Merge everything together
+        $guzzleConfig = ArrayHelper::merge($defaultConfig, $guzzleConfig, $config);
 
         return new Client($guzzleConfig);
     }